+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router';
-import Header from './components/header/Header';
+import { Header } from './components';
 
 function App() {
   return (
