+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { default as Card } from './card/Card';
+export { default as CardsList } from './cardsList/CardsList';
+export { default as Header } from './header/Header';
+export { default as Wrapper } from './wrapper/Wrapper';
+export { default as Search } from './search/Search';
