+++ src/pages/Main.tsx
@@ -1,7 +1,5 @@
-import Wrapper from '../components/wrapper/Wrapper';
 import React from 'react';
-import Search from '../components/search/Search';
-import CardsList from '../components/cardsList/CardsList';
+import { Wrapper, Search, CardsList } from '../components';
 
 export default class Main extends React.Component {
   render() {
